Add resource price endpoint

+++ src/main/java/com/albionrmtempire/datatransferobject/preprocessedorders/ItemOrderRequest.java
@@ -10,6 +10,6 @@ public class ItemOrderRequest extends PreProcessedOrder {
 
     public ItemOrderRequest(OrderRequest request) {
         super(request);
-        this.quality = request.quality();
+        this.quality = (short) request.quality();
     }
 }
